Move context extractor
