+++ src/auth/ctx.rs
@@ -1,3 +1,8 @@
+use axum::{
+	extract::FromRequestParts,
+	http::request::Parts,
+};
+
 #[derive(Clone, Debug)]
 pub struct Ctx {
 	user_id: i32,
@@ -11,6 +16,19 @@ pub struct Ctx {
 impl Ctx {
 	pub fn new(user_id: i32, username: String, firstname: String, surname: String, role: String) -> Self {
 		Self { user_id, username, firstname, surname, role }
+	}
+}
+
+impl<S: Send + Sync> FromRequestParts<S> for Ctx {
+	type Rejection = String;
+
+	async fn from_request_parts(parts: &mut Parts, _state: &S) -> Result<Self, String> {
+		match parts
+			.extensions
+			.get::<Ctx>() {
+				Some(ctx) => Ok(ctx.clone()),
+				None => Err("Missing user ctx".to_string()),
+			}
 	}
 }
